requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
APScheduler==3.10.4
nltk==3.8.1
requests==2.31.0 